**Improved Commit Message:**
**Journal Date Handling and Dashboard Enhancements:**

* Journal entry dates are now stored and sent in a list format instead of within an object.
* Implemented `getJournalEntries` API call in `api.js` to fetch the list of dates with journal entries.
* Separated calendar modal and memory card logic into dedicated components within `dashboard.jsx` for better organization.
* Fixed navigation issues from the memory page and the dashboard calendar.
