+++ frontend/src/pages/Dashboard.jsx
@@ -1,13 +1,12 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { motion } from "framer-motion";
-import Calendar from "react-calendar";
-import { format } from "date-fns";
 import { BookOpenIcon, CalendarIcon } from "@heroicons/react/24/outline";
 import MemorySearch from "../components/MemorySearch";
 import { useAuth } from "../context/AuthContext";
 import "react-calendar/dist/Calendar.css";
-
+import MemoryCard from "../components/MemoryCard";
+import CalendarModal from "../components/CalendarModal";
+import { getJournalEntryDates } from "../api";
 // Dummy data for demonstration
 const dummyMemories = [
   {
@@ -34,63 +33,18 @@ const dummyMemories = [
 export default function Dashboard() {
   const navigate = useNavigate();
   const [showCalendar, setShowCalendar] = useState(false);
-  const [selectedDate, setSelectedDate] = useState(null);
   const [searchQuery, setSearchQuery] = useState("");
+  const [journalDates, setJournalDates] = useState([]);
   const { user } = useAuth();
 
   // Function to check if a date has a journal entry
-  const hasJournal = (date) => {
-    const formattedDate = format(date, "yyyy-MM-dd");
-    return dummyMemories.some((memory) => memory.date === formattedDate);
+  useEffect(() => {
+    fetchDashboardData();
+  }, []);
+  const fetchDashboardData = async () => {
+    const dates = await getJournalEntryDates();
+    setJournalDates(dates);
   };
-
-  // Custom tile class for the calendar
-  const tileClassName = ({ date, view }) => {
-    if (view === "month") {
-      // Disable future dates
-      if (date > new Date()) {
-        return "opacity-25 cursor-not-allowed";
-      }
-      // Highlight dates with journal entries
-      if (hasJournal(date)) {
-        return "bg-blue-100 text-blue-800 font-medium hover:bg-blue-200";
-      }
-      // Style for dates without entries
-      return "text-gray-400";
-    }
-  };
-
-  // Disable tile function
-  const tileDisabled = ({ date }) => {
-    return date > new Date() || !hasJournal(date);
-  };
-
-  const MemoryCard = ({ memory }) => (
-    <motion.div
-      whileHover={{ scale: 1.02 }}
-      className="bg-white rounded-xl shadow-sm overflow-hidden cursor-pointer hover:shadow-md transition-shadow"
-      onClick={() => navigate(`/memory/${memory.id}`)}
-    >
-      <img
-        src={memory.thumbnail}
-        alt={memory.title}
-        className="w-full h-48 object-cover"
-      />
-      <div className="p-4">
-        <div className="flex justify-between items-start mb-2">
-          <h3 className="font-semibold text-lg text-gray-800">
-            {memory.title}
-          </h3>
-          <span className="text-2xl">{memory.mood}</span>
-        </div>
-        <p className="text-gray-600 text-sm mb-2">{memory.preview}</p>
-        <span className="text-sm text-gray-500">
-          {format(new Date(memory.date), "MMM d, yyyy")}
-        </span>
-      </div>
-    </motion.div>
-  );
-
   return (
     <div className="min-h-screen bg-gray-50">
       {/* Main Content */}
@@ -145,55 +99,13 @@ export default function Dashboard() {
           searchQuery={searchQuery}
           setSearchQuery={setSearchQuery}
           setShowCalendar={setShowCalendar}
-          navigate={navigate}
         />
         {/* Calendar Modal */}
         {showCalendar && (
-          <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center p-4 z-50">
-            <div className="bg-white rounded-xl p-6 max-w-md w-full">
-              <h3 className="text-lg font-semibold mb-4">Select a Date</h3>
-              <Calendar
-                onChange={setSelectedDate}
-                value={selectedDate}
-                tileClassName={tileClassName}
-                tileDisabled={tileDisabled}
-                maxDate={new Date()}
-                minDetail="month"
-                className="w-full border-0 shadow-sm rounded-lg"
-                navigationLabel={({ date }) => format(date, "MMMM yyyy")}
-                formatShortWeekday={(locale, date) => format(date, "EEE")}
-              />
-              <div className="mt-4 flex justify-end gap-3">
-                <button
-                  onClick={() => {
-                    setSelectedDate(null);
-                    setShowCalendar(false);
-                  }}
-                  className="px-4 py-2 bg-gray-100 text-gray-700 rounded-lg hover:bg-gray-200 transition-colors"
-                >
-                  Cancel
-                </button>
-                <button
-                  onClick={() => {
-                    if (selectedDate && hasJournal(selectedDate)) {
-                      setShowCalendar(false);
-                      navigate(
-                        `/memories/${format(selectedDate, "yyyy-MM-dd")}`
-                      );
-                    }
-                  }}
-                  disabled={!selectedDate || !hasJournal(selectedDate)}
-                  className={`px-4 py-2 rounded-lg transition-colors ${
-                    selectedDate && hasJournal(selectedDate)
-                      ? "bg-blue-600 text-white hover:bg-blue-700"
-                      : "bg-gray-300 text-gray-500 cursor-not-allowed"
-                  }`}
-                >
-                  View Journal
-                </button>
-              </div>
-            </div>
-          </div>
+          <CalendarModal
+            setShowCalendar={setShowCalendar}
+            journalDates={journalDates}
+          />
         )}
 
         {/* Recent Memories */}
